+++ changelog.md
@@ -1,2 +1,3 @@
 - Added OpenHUD integration
 - Fixed player not rendering in death history screen
+- Added "Add Waypoint" button in death history screen if OpenHUD is installed